snd_unit: generate open method for derived objects
In libhinawa, HinawaSndUnit has open method (=hinawa_snd_unit_open()) and
derived objects override it by own method. However the overridden methods
are missing from API crate as of hinawa-rs v0.1.0.

Affected objects and methods are:

    HinawaSndDg00x (hinawa_snd_dg00x_open)
    HinawaSndDice (hinawa_snd_dice_open)
    HinawaSndEfw (hinawa_snd_efw_open)
    HinawaSndMotu (hinawa_snd_motu_open)
    HinawaSndTscm (hinawa_snd_tscm_open)

This is due to the design of gtk-rs/gir. According to issue #139, when
derived objects override parent method, binding is not generated for the
methods to generate binding for the parent method. Therefore it's expected
not to generate bindings for the above methods.

This commit fixes the bug. A configuration is added to suppress binding
generation for the parent method. Then the overridden methods are
generated for derived object automatically in each trait. A manual binding
for the parent method is added by hand in manual trait. Users can decide
which open method is used by selecting the traits.

Reference: https://github.com/gtk-rs/gtk/pull/425
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>