add Cargo Workspace support
Cargo has a feature called as Workspace to aggregate several crates under
control. The feature is useful for this project because several FFI/Safe
crates are already under control.

This commit adds Workspace support. All of FFI/Safe crates are under the
additional workspace. All of Safe crates are handled as a default due to
test failure for FFI crates.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>