download glib related crates from crates.io instead of github.com
When under development, glib related crates are downloaded from
official repository in github.com and HEAD is used. They are nightly
and better for development. For published version, usage of crates
in crates.io is preferable.

This commit replaces dependency directives to download the crates from
crates.io. The crates compatible to "0.10" version are used.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>